Merged PR 118540: Hackfix: Exiting fullscreen once reduces ball launch delay
When running VG (any product) there is a delay between physical ball launch and when we see it on the screen. It's fixable by manually pressing alt+enter while playing, which changes the fullscreen mode to windowed. This fix persists even after going back to fullscreen. We don't know why this is happening.

This hackfix temporarily fixes the delay issue. It needs to be thoughly tested in different screens configuration.